+++ packages/host/modules/clusterio/impl.lua
@@ -1,10 +1,15 @@
 local api = require('modules/clusterio/api')
+local compat = require("modules/clusterio/compat")
 
+--- @class (exact) EventData.on_server_startup:EventData
 
 local function check_patch()
-	if global.clusterio_patch_number ~= clusterio_patch_number then
-		global.clusterio_patch_number = clusterio_patch_number
-		script.raise_event(api.events.on_server_startup, {})
+	local script_data = compat.script_data()
+	if script_data.clusterio_patch_number ~= clusterio_patch_number then
+		script_data.clusterio_patch_number = clusterio_patch_number
+		script.raise_event(api.events.on_server_startup, {
+			name = api.events.on_server_startup, tick = game.tick
+		})
 	end
 end
 
@@ -14,8 +19,9 @@ impl.events = {}
 impl.events[defines.events.on_tick] = check_patch
 
 impl.events[api.events.on_server_startup] = function()
-	if not global.clusterio then
-		global.clusterio = {
+	local script_data = compat.script_data()
+	if not script_data.clusterio then
+		script_data.clusterio = {
 			instance_id = nil,
 			instance_name = nil,
 		}
@@ -49,9 +55,10 @@ end
 -- Internal API
 clusterio_private = {}
 function clusterio_private.update_instance(new_id, new_name)
+	local script_data = compat.script_data()
 	check_patch()
-	global.clusterio.instance_id = new_id
-	global.clusterio.instance_name = new_name
+	script_data.clusterio.instance_id = new_id
+	script_data.clusterio.instance_name = new_name
 	script.raise_event(api.events.on_instance_updated, {
 		instance_id = new_id,
 		instance_name = new_name,
@@ -71,16 +78,17 @@ remote.add_interface('clusterio_api', {
 	end,
 
 	get_instance_id = function()
-		return global.clusterio.instance_id
+		return compat.script_data().clusterio.instance_id
 	end,
 
 	get_instance_name = function()
-		return global.clusterio.instance_name
+		return compat.script_data().clusterio.instance_name
 	end,
 
 	get_file_no = function()
-		global.clusterio_file_no = (global.clusterio_file_no or 0) + 1
-		return global.clusterio_file_no
+		local script_data = compat.script_data()
+		script_data.clusterio_file_no = (script_data.clusterio_file_no or 0) + 1
+		return script_data.clusterio_file_no
 	end,
 })