+++ plugins/research_sync/module/sync.lua
@@ -1,53 +1,79 @@
 local clusterio_api = require("modules/clusterio/api")
+local compat = require("modules/clusterio/compat")
+
+local v2_tech_progress = compat.version_ge("2.0.0")
 
 local sync = {}
 
-
+--- @param tech LuaTechnology
 local function get_technology_progress(tech)
 	if tech == tech.force.current_research then
 		return tech.force.research_progress
+	elseif v2_tech_progress then
+		return tech.saved_progress
 	else
+		--- @diagnostic disable-next-line
 		return tech.force.get_saved_technology_progress(tech.name)
 	end
 end
 
+--- @param tech LuaTechnology
+--- @param progress number
 local function set_technology_progress(tech, progress)
 	if tech == tech.force.current_research then
 		tech.force.research_progress = progress
+	elseif v2_tech_progress then
+		tech.saved_progress = progress
 	else
+		--- @diagnostic disable-next-line
 		tech.force.set_saved_technology_progress(tech.name, progress)
 	end
 end
 
-sync.events = {}
-sync.events[clusterio_api.events.on_server_startup] = function(event)
-	if not global.research_sync then
-		global.research_sync = {
-			technologies = {},
-		}
+--- @param no_early_return boolean?
+--- @return table
+local function get_script_data(no_early_return)
+	local script_data = compat.script_data()
+	local research_sync = script_data.research_sync
+	if research_sync and not no_early_return then
+		return research_sync
 	end
 
-	-- Used when syncing completed technologies from the controller
-	global.research_sync.ignore_research_finished = false
+	research_sync = research_sync or {}
+	research_sync.technologies = research_sync.technologies or {}
+
+	if research_sync.ignore_research_finished == nil then
+		research_sync.ignore_research_finished = false
+	end
 
 	local force = game.forces["player"]
 	for _, tech in pairs(force.technologies) do
 		local progress = get_technology_progress(tech)
-		global.research_sync.technologies[tech.name] = {
+		research_sync.technologies[tech.name] = {
 			level = tech.level,
 			researched = tech.researched,
 			progress = progress,
 		}
 	end
+
+	script_data.research_sync = research_sync
+	return research_sync
 end
 
+sync.events = {}
+sync.events[clusterio_api.events.on_server_startup] = function(event)
+	get_script_data(true)
+end
+
+--- @param tech LuaTechnology
 local function get_contribution(tech)
 	local progress = get_technology_progress(tech)
 	if not progress then
 		return 0, nil
 	end
 
-	local prev_tech = global.research_sync.technologies[tech.name]
+	local script_data = get_script_data()
+	local prev_tech = script_data.technologies[tech.name]
 	if prev_tech.progress and prev_tech.level == tech.level then
 		return progress - prev_tech.progress, progress
 	else
@@ -55,6 +81,7 @@ local function get_contribution(tech)
 	end
 end
 
+--- @param tech LuaTechnology
 local function send_contribution(tech)
 	local contribution, progress = get_contribution(tech)
 	if contribution ~= 0 then
@@ -63,10 +90,12 @@ local function send_contribution(tech)
 			level = tech.level,
 			contribution = contribution,
 		})
-		global.research_sync.technologies[tech.name].progress = progress
+		local script_data = get_script_data()
+		script_data.technologies[tech.name].progress = progress
 	end
 end
 
+--- @param event EventData.on_research_started
 sync.events[defines.events.on_research_started] = function(event)
 	local tech = event.last_research
 	if tech then
@@ -74,13 +103,15 @@ sync.events[defines.events.on_research_started] = function(event)
 	end
 end
 
+--- @param event EventData.on_research_finished
 sync.events[defines.events.on_research_finished] = function(event)
-	if global.research_sync.ignore_research_finished then
+	local script_data = get_script_data()
+	if research_sync.ignore_research_finished then
 		return
 	end
 
 	local tech = event.research
-	global.research_sync.technologies[tech.name] = {
+	script_data.technologies[tech.name] = {
 		level = tech.level,
 		researched = tech.researched,
 	}
@@ -121,10 +152,11 @@ function research_sync.dump_technologies()
 	if #techs == 0 then
 		rcon.print("[]")
 	else
-		rcon.print(game.table_to_json(techs))
+		rcon.print(compat.table_to_json(techs))
 	end
 end
 
+--- @param data string
 function research_sync.sync_technologies(data)
 	local force = game.forces["player"]
 
@@ -133,8 +165,9 @@ function research_sync.sync_technologies(data)
 	local progressIndex = 3
 	local researchedIndex = 4
 
-	global.research_sync.ignore_research_finished = true
-	for _, tech_data in pairs(game.json_to_table(data)) do
+	local script_data = get_script_data()
+	script_data.ignore_research_finished = true
+	for _, tech_data in pairs(compat.json_to_table(data) --[[@as table]]) do
 		local tech = force.technologies[tech_data[nameIndex]]
 		if tech and tech.level <= tech_data[levelIndex] then
 			local new_level = math.min(tech_data[levelIndex], tech.prototype.max_level)
@@ -162,18 +195,20 @@ function research_sync.sync_technologies(data)
 				progress = get_technology_progress(tech)
 			end
 
-			global.research_sync.technologies[tech.name] = {
+			script_data.technologies[tech.name] = {
 				level = tech.level,
 				researched = tech.researched,
 				progress = progress,
 			}
 		end
 	end
-	global.research_sync.ignore_research_finished = false
+	script_data.ignore_research_finished = false
 end
 
+--- @param data string
 function research_sync.update_progress(data)
-	local techs = game.json_to_table(data)
+	local script_data = get_script_data()
+	local techs = compat.json_to_table(data) --[[@as table]]
 	local force = game.forces["player"]
 
 	for _, controllerTech in ipairs(techs) do
@@ -181,7 +216,7 @@ function research_sync.update_progress(data)
 		if tech and tech.level == controllerTech.level then
 			send_contribution(tech)
 			set_technology_progress(tech, controllerTech.progress)
-			global.research_sync.technologies[tech.name] = {
+			script_data.technologies[tech.name] = {
 				level = tech.level,
 				progress = controllerTech.progress
 			}
@@ -189,6 +224,8 @@ function research_sync.update_progress(data)
 	end
 end
 
+--- @param name string
+--- @param level number
 function research_sync.research_technology(name, level)
 	local force = game.forces["player"]
 	local tech = force.technologies[name]
@@ -200,7 +237,8 @@ function research_sync.research_technology(name, level)
 		level = tech.prototype.max_level
 	end
 
-	global.research_sync.ignore_research_finished = true
+	local script_data = get_script_data()
+	script_data.ignore_research_finished = true
 	if tech == force.current_research and tech.level == level then
 		force.research_progress = 1
 
@@ -215,9 +253,9 @@ function research_sync.research_technology(name, level)
 		end
 		game.play_sound { path = "utility/research_completed" }
 	end
-	global.research_sync.ignore_research_finished = false
+	script_data.ignore_research_finished = false
 
-	global.research_sync.technologies[tech.name] = {
+	script_data.technologies[tech.name] = {
 		level = tech.level,
 		researched = tech.researched,
 	}